+++ lib/puffer_pages/backends/models/layout.rb
@@ -3,6 +3,4 @@ class PufferPages::Layout < ActiveRecord::Base
 
   validates_presence_of :name
   validates_uniqueness_of :name
-
-  has_many :pages, :class_name => '::Page'
 end