+++ lib/puffer_pages/renderer.rb
@@ -5,7 +5,7 @@ module PufferPages
       puffer_page = puffer_pages_template(options[:action].presence || options[:file]) unless puffer_page
       @view.assign(@view.assigns.merge!(
         'puffer_page' => puffer_page,
-        'puffer_page_drops' => puffer_page_drops
+        'puffer_page_render_options' => puffer_page_render_options
       ))
 
       super
@@ -15,8 +15,9 @@ module PufferPages
     end
 
     def find_layout(layout, keys)
-      layout = "<%= render :inline => @puffer_page.render(default_drops.merge(@puffer_page_drops)),
-        :layout => @puffer_page.layout_for_render %>"
+      layout = "<% @puffer_page_render_options[:drops].merge!(default_drops) %><%= render(
+        :inline => @puffer_page.render(@puffer_page_render_options),
+        :layout => @puffer_page.layout_for_render) %>"
 
       handler = ActionView::Template.handler_for_extension("erb")
       ActionView::Template.new(layout, "puffer pages layout wrapper", handler, :locals => keys)
@@ -26,11 +27,16 @@ module PufferPages
       ::Page.find_layout_page(suggest.presence || @view.request.path_info)
     end
 
-    def puffer_page_drops
-      @view.assigns.inject({}) do |drops, (key, value)|
-        drops[key] = value if value.respond_to?(:to_liquid)
-        drops
-      end.stringify_keys
+    def puffer_page_render_options
+      drops, registers = {}, {}
+      @view.assigns.each do |key, value|
+        if value.respond_to?(:to_liquid)
+          drops[key] = value
+        else
+          registers[key] = value
+        end
+      end
+      {:drops => drops, :registers => registers}
     end
   end
 end