+++ spec/lib/drops_spec.rb
@@ -2,14 +2,14 @@ require 'spec_helper'
 
 describe PufferPages::Liquid::PageDrop do
 
-  def render(current_page, page = nil)
+  def render current_page, page = nil
     current_page.render({
       'self' => current_page.to_drop(current_page, controller),
       'page' => (page.to_drop(current_page, controller) if page)
     })
   end
 
-  def render_layout(layout, current_page, page = nil)
+  def render_layout layout, current_page, page = nil
     current_page.render_layout(layout, {
       'self' => current_page.to_drop(current_page),
       'page' => (page.to_drop(current_page) if page)